pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tower-cli"
dynamic = ["version"]
description = "Permission evaluation agent for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "thedhanawada" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "InquirerPy>=0.3.4",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/thedhanawada/tower-cli"
Repository = "https://github.com/thedhanawada/tower-cli"
Issues = "https://github.com/thedhanawada/tower-cli/issues"

[project.scripts]
tower = "tower.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.dynamic]
version = {attr = "tower.__version__"}
